Remove all Lua-exposed bindings/usage of PositionOrUnit class
C++ does implicit conversion if a Unit or a Position is passed to a function that takes a PositionOrUnit, but when exposed to Lua, it explicitly requires a PositionOrUnit instance. To keep functionality the same, Lua needs overloads that take a Unit or a Position, which makes PositionOrUnit obsolete so it can be dropped entirely.

Also, certain Unit member function bindings have been split into separate files to keep memory consumption lower when compiling the Unit class. Note: Keeping PositionOrUnit functions bound and adding both Position and Unit overloads made memory consumption explode during compilation (seems like there's a practical limit to the number of overloads a function can have in sol2 before it becomes un-compilable due to ballooning memory usage)